+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Tutorial enhancements:
 - Enhanced ordering of sphinx gallery tutorials to use alphabetic ordering based on tutorial headings. @oruebel (#1399)
+- Updated the general tutorial to add documentation about the ``Images`` type. @bendichter (#1353)
 
 ## PyNWB 2.0.0 (August 13, 2021)